Update README.md
Wonder why you haven’t had much luck with digital marketing for your business?

You believe in getting 𝐑𝐎𝐈 in the short term. But wise digital marketing is for the long run.

You have sales mentality, and you are only interested  in quick returns. You see marketing as an expense and not an investment. Don’t understand brand building or long term value oriented thinking.

Our observations are based on the substantial clientele we have handled from various industries and the digital pain areas we have solved for each business. (Take a look at the two figures to understand the breadth and depth of our work.)

We have thus been able to filter the kind of leaders and CXOs who are serious about their organisation’s digital transformation and are committed to investing in the right digital marketing resources. The others are non-digital CXOs who do not even have the right questions to ask us!

We try our best to avoid onboarding businesses who do not respect the basic requirements for digital transformation. We saw a sales person is heading digital marketing and there is no dedicated marketing head.

Even if we work with you, you wouldn’t have the right human resources with their digital KRAs to understand the value we bring to your digital transformation. A dedicated marketing team (consisting of a marketing qualified team) suggests the marketing mindset of an organisation.

We’ll work with you if you qualify as per the following criteria:- 

   01. Growth or digital  mindset or acceptance of digital transformation 

   02. Shedding of digital denial attitude

   03. Willingness to truly understand business transformation 

   04. Patience to invest in brand discovery

   05.Willingness to invest in the right measures of digital marketing

If you wish to invest in better digital marketing and have the resources and patience for it, talk to us.


𝐆𝐞𝐭 𝐢𝐧 𝐭𝐨𝐮𝐜𝐡:  +𝟗𝟏 𝟗𝟗𝟔𝟕 𝟔𝟑𝟎𝟑 𝟐𝟗